requirements.txt
flask==3.0.0
flask-cors==4.0.0
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
matplotlib==3.8.2
gunicorn==21.2.0
